requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
peewee==3.17.8
python-dotenv==1.0.1
flasgger==0.9.7.1
pydantic==2.7.1
uvicorn==0.32.1
psycopg2-binary==2.9.9
requests==2.32.3